+++ example/integration_test/src/tests/unit/src/view/test_app_menu.dart
@@ -9,5 +9,74 @@ import '_unit_test_view.dart';
 
 void testAppMenu(WidgetTester tester) {
   //
-//  test('Test AppMenu', () async {});
+//  test('Test AppMenu', () async {
+  final menu = AppMenu();
+
+  menu.items;
+
+  menu.menuEntries;
+
+  menu.itemBuilder;
+
+  menu.initialValue;
+
+  menu.inOpened;
+
+  menu.inSelected;
+
+  menu.inCanceled;
+
+  menu.tooltip;
+
+  menu.elevation;
+
+  menu.padding;
+
+  menu.splashRadius;
+
+  menu.child;
+
+  menu.icon;
+
+  menu.offset;
+
+  menu.enabled;
+
+  menu.shape;
+
+  menu.color;
+
+  menu.enableFeedback;
+
+  menu.iconSize;
+
+  menu.constraints;
+
+  menu.position;
+
+  menu.clipBehavior;
+
+  menu.context;
+
+  menu.setMenu();
+
+  menu.showButtonMenu();
+
+  menu.onMenuEntries();
+
+  menu.onItemBuilder();
+
+//  menu.onSelected();
+
+  menu.onCanceled();
+
+  menu.onShape();
+
+  menu.onPosition();
+
+  final con = AppMenuController();
+
+  con.onCanceled();
+
+//  });
 }