+++ example/integration_test/src/tests/unit/unit_tests.dart
@@ -21,7 +21,7 @@ unitTests(WidgetTester tester) async {
 
   testLoadingScreen(tester);
 
-  testAppView(tester);
+  testAppObject(tester);
 
   testAppMenu(tester);