+++ lib/view/rx_core/_view_rx.dart
@@ -1,9 +1,0 @@
-///
-
-export 'rx_impl.dart';
-
-export 'rx_bool.dart';
-
-export 'rx_num.dart';
-
-export 'rx_string.dart';